+++ cmd/livesim2/app/timesubs_test.go
@@ -252,7 +252,7 @@ func TestTimeSubsInitSegment(t *testing.T) {
 		TimeoutS:  0,
 		LogFormat: logging.LogDiscard,
 	}
-	_, err := logging.InitZerolog(cfg.LogLevel, cfg.LogFormat)
+	err := logging.InitSlog(cfg.LogLevel, cfg.LogFormat)
 	require.NoError(t, err)
 	server, err := SetupServer(context.Background(), &cfg)
 	require.NoError(t, err)
@@ -343,7 +343,7 @@ func TestTimeSubsMediaSegment(t *testing.T) {
 		TimeoutS:  0,
 		LogFormat: logging.LogDiscard,
 	}
-	_, err := logging.InitZerolog(cfg.LogLevel, cfg.LogFormat)
+	err := logging.InitSlog(cfg.LogLevel, cfg.LogFormat)
 	require.NoError(t, err)
 	server, err := SetupServer(context.Background(), &cfg)
 	require.NoError(t, err)